% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTwofoldNs.R
\docType{data}
\name{dataTwofoldNS}
\alias{dataTwofoldNS}
\title{Simulated dataset Under Two Fold Subarea level model with Normal distribution and Non-sampled subarea}
\format{
A data frame with 90 row and 7 column:
\describe{
 \item{y}{Direct estimation of subarea mean \eqn{y}}
 \item{x1}{Auxiliary variabel of \eqn{x_1}}
 \item{x2}{Auxiliary variabel of \eqn{x_2}}
 \item{codearea}{Index that describes the code relating to area for each subarea}
 \item{w}{Unit proportion on each subarea or weight}
 \item{vardir}{Sampling variance of y}
}
}
\usage{
dataTwofoldNS
}
\description{
\enumerate{
  \item A dataset to simulate Small Area Estimation using Hierarchical Bayesian method under Two Fold Subarea level model with Normal distribution and Non-sampled subarea
  \item This data contains NA values that indicates no sampled at one or more Subareas. It uses the \code{\link{dataTwofold}} with the direct estimates and the related variances in 10 subareas are missing.
 }
}
\keyword{datasets}
